requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
celery
redis
asyncpg
openai
scipy
scikit-learn
python-dateutil
pytest-asyncio
httpx
beautifulsoup4
trafilatura
lxml
APScheduler
reportlab
openai
tenacity
argon2-cffi

